refactor(gpu): thread log2 values as u32 through the GPU stack
Follow-up to typing get_domain_size_log2() as u32: eliminate the remaining
u32 → usize → u32 roundtrips by making log2-semantic values u32 end-to-end
inside the GPU stack — ~19 production/test-helper params (final_trace_size_
log_2, trace_len_log2, query_domain_log2, initial/final_trace_log_2 across
proof orchestration, whir fold scheduling, gkr forward/backward, pow), the
execution_prover origin literal, and the test-side TRACE_LEN/FINAL_TRACE
consts and locals.

Casts now appear only at genuine boundaries: upstream CPU APIs (cs compiler,
example_configs, dimension-reduction reference evaluator, transcript utils),
upstream ProverConfig usize fields, gpu_gkr_model's host-side layout API,
and slice indexing/allocation sizes. Deleted intra-stack shims include the
whir/mod.rs trailing_zeros-then-recast chain (9 casts), the dim-reducing
tower LOG_BLOCK roundtrip, the fold-schedule query-domain computations, and
the trace-holder log_domain_size assert casts.

Values are provably unchanged (widening-only at boundaries; all log2s ≤ 30).
Validated: lib + test-target + gpu_execution_prover/gpu_program_prover
compile; circuit_type unit tests; run_jump_branch_slt_proof_parity_test
end-to-end on GPU (bit-identical proof vs CPU).

+++ gpu/circuit_prover/src/prover/gkr/forward/tests/helpers.rs
@@ -124,8 +124,8 @@ pub(super) fn attach_test_dim_reducing_tower_layout(
     storage: &mut GpuGKRStorage<BF, E4>,
     initial_layer_idx: usize,
     initial_output_map: &BTreeMap<OutputType, Vec<GKRAddress>>,
-    initial_trace_log_2: usize,
-    final_trace_log_2: usize,
+    initial_trace_log_2: u32,
+    final_trace_log_2: u32,
 ) {
     use crate::prover::gkr::gkr_address_audit::AddressClass;
     use crate::prover::gkr::storage_layout::{
@@ -134,7 +134,7 @@ pub(super) fn attach_test_dim_reducing_tower_layout(
 
     let trace_len = 1usize << initial_trace_log_2;
     let total_rounds = initial_trace_log_2.saturating_sub(final_trace_log_2);
-    let total_layers = initial_layer_idx + total_rounds + 1;
+    let total_layers = initial_layer_idx + total_rounds as usize + 1;
     let mut layers = vec![GpuGKRLayerLayout::default(); total_layers];
 
     // Register the test fixture inputs at `initial_layer_idx`. The test
